Keep a list of domains that have been blocked.
Store a list of domains which we have blocked, encoded in JSON, in the
key-value store.  The list is displayed on the 'blocked' tab in the
extension config page.

If a domain is re-enabled, it is no longer displayed on the 'blocked'
tab, and it will be removed from the stored list the next time it is
updated.