Make blocked domain list items links.
Make the items in the blocked domains links to mail settings page for
that domain.  This required a major reworking of how they are stored,
they are now stored as a map from site_id to domain name instead of a
simple list of domains.

I also changed the way the domain status is looked up, so there is only
one XML-RPC call, instead of one per blocked domain.